requirements.txt
flask==3.0.0
flask-cors==4.0.0
supabase==2.3.0
python-dotenv==1.0.0
requests==2.31.0
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gunicorn==21.2.0
